Extract SignerProvider query logic into SignerProviderQuery helper
Move the signing/encryption logic out of the SignerProvider ContentProvider
into a reusable SignerProviderQuery helper. SignerProvider now just adapts
the binder call (context + calling package) and delegates.

The NIP-46 relay path (EventNotificationConsumer) previously reached this
logic through a same-app ContentResolver.query() round-trip; it now calls
SignerProviderQuery.query() directly, avoiding the IPC.

The helper's parameters are named for intent instead of mirroring the
ContentProvider signature, and the sortOrder-based package-name override is
removed. The requester identity is now always derived by Amber (the calling
package for IPC, the client pubkey for NIP-46) and never read from
caller-supplied query arguments, so external apps can no longer impersonate
another requester.
